requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
numpy==1.24.4
websockets==11.0.3
pydantic==2.5.0
python-multipart==0.0.6
qiskit==0.44.1
qiskit-ibm-runtime==0.15.1
qiskit-ibm-provider==0.7.2
python-dotenv==1.0.0
cryptography==41.0.4
asyncio
flask
gunicorn
qiskit
numpy
scipy

